+++ src/style/load_iconset.ts
@@ -1,6 +1,6 @@
 import {PbfReader} from 'pbf';
 import {getArrayBuffer, ResourceType} from "../util/ajax";
-import {readIconSet, type Icon} from "../data/usvg/usvg_pb_decoder";
+import {readIconSetLazy, type Icon} from "../data/usvg/usvg_pb_decoder";
 import browser from '../util/browser';
 
 import type {Callback} from "../types/callback";
@@ -45,7 +45,7 @@ export async function loadIconset(
         const {data} = await getArrayBuffer(request, signal);
         const result: StyleImages = {};
 
-        const iconSet = readIconSet(new PbfReader(data));
+        const iconSet = readIconSetLazy(new PbfReader(data));
 
         for (const icon of iconSet.icons) {
             const styleImage: StyleImage = {